requirements.txt
boto==2.39.0
file-magic==0.3.0
pathlib==1.0.1
wheel==0.24.0
